+++ frontend-user/src/app/companies/[id]/page.tsx
@@ -2,7 +2,6 @@
 import { useParams } from "next/navigation";
 import { useEffect, useState } from "react";
 import Image from "next/image";
-import Link from "next/link";
 import { Company } from "../../../types/company"; // 型定義ファイルをインポート
 import { Review } from "../../../types/review"; // 型定義ファイルをインポート
 import { Recruitment } from "../../../types/recruitment"; // 型定義ファイルをインポート
@@ -18,6 +17,7 @@ interface CompanyWithReviews {
 const CompanyDetailPage = () => {
   const { id } = useParams();
   const [data, setData] = useState<CompanyWithReviews | null>(null);
+  const [refreshTrigger, setRefreshTrigger] = useState(false);
   const [tabIndex, setTabIndex] = useState(0);
 
   useEffect(() => {
@@ -27,7 +27,7 @@ const CompanyDetailPage = () => {
         .then((data) => setData(data))
         .catch((error) => console.error("Error fetching company:", error));
     }
-  }, [id]);
+  }, [id, refreshTrigger]);
 
   if (!data) {
     return <p>Loading...</p>;
@@ -40,12 +40,6 @@ const CompanyDetailPage = () => {
       {/* Company Header */}
       <div className="flex items-center p-4 text-white bg-main-col rounded-lg">
         <div className="flex-grow text-2xl font-bold">{`${company.name}`}</div>
-        <Link
-          href={`/companies/${company.id}/reviews/new`}
-          className="bg-sub-col text-white font-bold px-4 py-2 rounded"
-        >
-          口コミ投稿
-        </Link>
       </div>
       <div className="w-full mb-6">
         <Image
@@ -59,7 +53,13 @@ const CompanyDetailPage = () => {
       <p className="text-sm text-main-col font-bold">{company.description}</p>
 
       <TabContext.Provider
-        value={{ tabIndex, setTabIndex, recruitments, reviews }}
+        value={{
+          tabIndex,
+          setTabIndex,
+          setRefreshTrigger,
+          recruitments,
+          reviews,
+        }}
       >
         <Contents />
       </TabContext.Provider>
